chore: Newton line search deleted; divergence defers to the controller
- The Newton solver has a single acceptance scheme: the NLNewton
  update-error bound. A diverging or stagnant solve exits with a
  nonzero status and the adaptive step controller rejects the step
  and shrinks dt; there is no in-solve residual-descent line search.
- newton_damping, newton_max_backtracks, residual_temp and
  stage_base_bt buffers (and their location parameters), and the
  NEWTON_BACKTRACKING_NO_SUITABLE_STEP status code no longer exist.
- The Newton norm chain carries one device function: DIRK/FIRK
  correction norms build only the stage-scaled correction norm,
  stored as the factory's single cached output; the
  correction_norm_function config field and CorrectionNormCache are
  gone, and NewtonKrylov.update defers wholly to the parent's
  norm_device_function refresh.
- prev_theta_location and krylov_iters_local_location are exposed
  through the settings chain alongside the other buffer locations.
- The exit status is two selects: divergence-vs-max-iters, OR'd with
  the last linear status, gated on convergence.
- Instrumented mirrors and the CPU reference implement the identical
  flow; the newton_iteration_scale logging column is removed
  end-to-end.
- Newton convergence edge cases run through layered session-scoped
  conftest fixtures (case -> residual/operator system -> solver ->
  kernel), compiling once per parameter set.
