+++ docs/source/API_reference/integrators/algorithms.rst
@@ -136,10 +136,6 @@ the Krylov and preconditioner settings below apply to it and the
      - ``1e-6``
    * - ``newton_max_iters``
      - ``100``
-   * - ``newton_damping``
-     - ``0.5``
-   * - ``newton_max_backtracks``
-     - ``0``
    * - ``krylov_atol`` / ``krylov_rtol``
      - ``1e-6``
    * - ``krylov_max_iters``
