+++ docs/source/user_guide/configuration.rst
@@ -160,7 +160,6 @@ below may be passed directly to :func:`~cubie.solve_ivp`,
      - :doc:`optional_arguments`
    * - Algorithm
      - ``newton_atol``, ``newton_rtol``, ``newton_max_iters``,
-       ``newton_damping``, ``newton_max_backtracks``,
        ``krylov_atol``, ``krylov_rtol``, ``krylov_max_iters``,
        ``linear_correction_type``, ``preconditioner_type``,
        ``preconditioner_order``
