+++ tests/conftest.py
@@ -495,8 +495,6 @@ def solver_settings(solver_settings_override, system, precision):
         "preconditioner_order": 2,
         "krylov_max_iters": 50,
         "newton_max_iters": 50,
-        "newton_damping": precision(0.85),
-        "newton_max_backtracks": 25,
         "min_gain": precision(0.1),
         "max_gain": precision(5.0),
         "safety": precision(0.9),
@@ -526,7 +524,6 @@ def solver_settings(solver_settings_override, system, precision):
         "krylov_rtol",
         "newton_atol",
         "newton_rtol",
-        "newton_damping",
         "kp",
         "ki",
         "kd",
