+++ tests/integrators/algorithms/instrumented/backwards_euler_predict_correct.py
@@ -119,7 +119,6 @@ class InstrumentedBackwardsEulerPCStep(InstrumentedBackwardsEulerStep):
             newton_iteration_guesses,
             newton_residuals,
             newton_squared_norms,
-            newton_iteration_scale,
             linear_initial_guesses,
             linear_iteration_guesses,
             linear_residuals,
@@ -233,7 +232,6 @@ class InstrumentedBackwardsEulerPCStep(InstrumentedBackwardsEulerStep):
                 newton_iteration_guesses,
                 newton_residuals,
                 newton_squared_norms,
-                newton_iteration_scale,
                 linear_initial_guesses,
                 linear_iteration_guesses,
                 linear_residuals,
