+++ tests/integrators/algorithms/instrumented/generic_dirk.py
@@ -43,8 +43,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
         newton_atol: Optional[float] = None,
         newton_rtol: Optional[float] = None,
         newton_max_iters: Optional[int] = None,
-        newton_damping: Optional[float] = None,
-        newton_max_backtracks: Optional[int] = None,
         tableau: DIRKTableau = DEFAULT_DIRK_TABLEAU,
         n_drivers: int = 0,
         stage_increment_location: Optional[str] = None,
@@ -99,12 +97,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
         newton_max_iters
             Maximum iterations permitted for the Newton solver. If None, uses
             default from NewtonKrylovConfig.
-        newton_damping
-            Damping factor applied within Newton updates. If None, uses
-            default from NewtonKrylovConfig.
-        newton_max_backtracks
-            Maximum number of backtracking steps within the Newton solver. If
-            None, uses default from NewtonKrylovConfig.
         tableau
             DIRK tableau describing the coefficients. Defaults to
             :data:`DEFAULT_DIRK_TABLEAU`.
@@ -185,10 +177,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
             solver_kwargs["newton_rtol"] = newton_rtol
         if newton_max_iters is not None:
             solver_kwargs["newton_max_iters"] = newton_max_iters
-        if newton_damping is not None:
-            solver_kwargs["newton_damping"] = newton_damping
-        if newton_max_backtracks is not None:
-            solver_kwargs["newton_max_backtracks"] = newton_max_backtracks
 
         # Call parent __init__ to create solver instances
         super().__init__(
@@ -429,7 +417,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
             newton_iteration_guesses,
             newton_residuals,
             newton_squared_norms,
-            newton_iteration_scale,
             linear_initial_guesses,
             linear_iteration_guesses,
             linear_residuals,
@@ -525,7 +512,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
                         newton_iteration_guesses,
                         newton_residuals,
                         newton_squared_norms,
-                        newton_iteration_scale,
                         linear_initial_guesses,
                         linear_iteration_guesses,
                         linear_residuals,
@@ -657,7 +643,6 @@ class InstrumentedDIRKStep(InstrumentedODEImplicitStep):
                         newton_iteration_guesses,
                         newton_residuals,
                         newton_squared_norms,
-                        newton_iteration_scale,
                         linear_initial_guesses,
                         linear_iteration_guesses,
                         linear_residuals,
