+++ tests/integrators/algorithms/instrumented/generic_erk.py
@@ -132,7 +132,6 @@ class InstrumentedERKStep(ERKStep):
             newton_iteration_guesses,
             newton_residuals,
             newton_squared_norms,
-            newton_iteration_scale,
             linear_initial_guesses,
             linear_iteration_guesses,
             linear_residuals,
