+++ tests/integrators/algorithms/instrumented/generic_firk.py
@@ -70,8 +70,6 @@ class InstrumentedFIRKStep(InstrumentedODEImplicitStep):
         newton_atol: Optional[float] = None,
         newton_rtol: Optional[float] = None,
         newton_max_iters: Optional[int] = None,
-        newton_damping: Optional[float] = None,
-        newton_max_backtracks: Optional[int] = None,
         tableau: FIRKTableau = DEFAULT_FIRK_TABLEAU,
         n_drivers: int = 0,
         stage_increment_location: Optional[str] = None,
@@ -126,12 +124,6 @@ class InstrumentedFIRKStep(InstrumentedODEImplicitStep):
         newton_max_iters
             Maximum iterations permitted for the Newton solver. If None, uses
             default from NewtonKrylovConfig.
-        newton_damping
-            Damping factor applied within Newton updates. If None, uses
-            default from NewtonKrylovConfig.
-        newton_max_backtracks
-            Maximum number of backtracking steps within the Newton solver. If
-            None, uses default from NewtonKrylovConfig.
         tableau
             FIRK tableau describing the coefficients. Defaults to
             :data:`DEFAULT_FIRK_TABLEAU`.
@@ -213,10 +205,6 @@ class InstrumentedFIRKStep(InstrumentedODEImplicitStep):
             solver_kwargs["newton_rtol"] = newton_rtol
         if newton_max_iters is not None:
             solver_kwargs["newton_max_iters"] = newton_max_iters
-        if newton_damping is not None:
-            solver_kwargs["newton_damping"] = newton_damping
-        if newton_max_backtracks is not None:
-            solver_kwargs["newton_max_backtracks"] = newton_max_backtracks
 
         newton_norm = FIRKCorrectionNorm(
             precision=precision,
@@ -445,7 +433,6 @@ class InstrumentedFIRKStep(InstrumentedODEImplicitStep):
             newton_iteration_guesses,
             newton_residuals,
             newton_squared_norms,
-            newton_iteration_scale,
             linear_initial_guesses,
             linear_iteration_guesses,
             linear_residuals,
@@ -513,7 +500,6 @@ class InstrumentedFIRKStep(InstrumentedODEImplicitStep):
                 newton_iteration_guesses,
                 newton_residuals,
                 newton_squared_norms,
-                newton_iteration_scale,
                 linear_initial_guesses,
                 linear_iteration_guesses,
                 linear_residuals,
