+++ tests/integrators/algorithms/instrumented/generic_rosenbrock_w.py
@@ -384,7 +384,6 @@ class InstrumentedRosenbrockWStep(InstrumentedODEImplicitStep):
             newton_iteration_guesses,
             newton_residuals,
             newton_squared_norms,
-            newton_iteration_scale,
             linear_initial_guesses,
             linear_iteration_guesses,
             linear_residuals,
