+++ tests/integrators/matrix_free_solvers/conftest.py
@@ -1,8 +1,213 @@
+from math import copysign as math_copysign
+
 import numpy as np
 import pytest
 from cubie.cuda_simsafe import cuda
 
+from cubie.integrators.matrix_free_solvers import CUBIE_RESULT_CODES
+from cubie.integrators.matrix_free_solvers.linear_solver import (
+    MRLinearSolver,
+)
+from cubie.integrators.matrix_free_solvers.newton_krylov import (
+    NewtonKrylov,
+)
 from cubie.odesystems.symbolic.symbolicODE import create_ODE_system
+
+# Each case runs two sequential solves in one thread sharing the
+# solver's persistent scratch, so warm-started contraction history
+# carries from the first solve into the second.
+NEWTON_CONVERGENCE_EDGE_CASES = {
+    "small-first-step": dict(
+        kind="zero",
+        newton_atol=1e-6,
+        newton_max_iters=4,
+        krylov_atol=1e-6,
+        krylov_max_iters=8,
+        initials=(3.0, 3.0),
+        expected_statuses=(
+            CUBIE_RESULT_CODES.SUCCESS,
+            CUBIE_RESULT_CODES.SUCCESS,
+        ),
+        expected_counts=(1, 1),
+        expected_finals=(3.0, 3.0),
+        final_tolerance=0.0,
+    ),
+    "warm-start": dict(
+        kind="linear",
+        newton_atol=1e-2,
+        newton_max_iters=8,
+        krylov_atol=1e-6,
+        krylov_max_iters=8,
+        initials=(3.0, 3.9999),
+        expected_statuses=(
+            CUBIE_RESULT_CODES.SUCCESS,
+            CUBIE_RESULT_CODES.SUCCESS,
+        ),
+        expected_counts=(2, 1),
+        expected_finals=(4.0, 4.0),
+        final_tolerance=1e-3,
+    ),
+    "stagnation-divergence": dict(
+        kind="constant",
+        newton_atol=1e-2,
+        newton_max_iters=4,
+        krylov_atol=1e-6,
+        krylov_max_iters=8,
+        initials=(0.0, 0.0),
+        expected_statuses=(
+            CUBIE_RESULT_CODES.NEWTON_DIVERGENCE,
+            CUBIE_RESULT_CODES.NEWTON_DIVERGENCE,
+        ),
+        expected_counts=(2, 2),
+        expected_finals=(-1.0, -1.0),
+        final_tolerance=1e-6,
+    ),
+    "theta-growth-divergence": dict(
+        kind="root",
+        newton_atol=1e-2,
+        newton_max_iters=4,
+        krylov_atol=1e-6,
+        krylov_max_iters=8,
+        initials=(1.0, 1.0),
+        expected_statuses=(
+            CUBIE_RESULT_CODES.NEWTON_DIVERGENCE,
+            CUBIE_RESULT_CODES.NEWTON_DIVERGENCE,
+        ),
+        expected_counts=(2, 2),
+        expected_finals=(-3.0, -3.0),
+        final_tolerance=1e-4,
+    ),
+}
+
+
+@pytest.fixture(scope="session")
+def newton_edge_case(request):
+    """Return one named Newton convergence edge case."""
+    return NEWTON_CONVERGENCE_EDGE_CASES[request.param]
+
+
+@pytest.fixture(scope="session")
+def newton_edge_system(newton_edge_case, precision):
+    """Compile the residual and operator for one edge case."""
+    kind = newton_edge_case["kind"]
+    target = precision(4.0)
+
+    @cuda.jit(device=True)
+    def residual(
+        state, parameters, drivers, t, h, a_ij, base_state, out
+    ):
+        if kind == "zero":
+            out[0] = precision(0.0)
+        elif kind == "linear":
+            out[0] = target - state[0]
+        elif kind == "constant":
+            out[0] = precision(1.0)
+        else:
+            magnitude = abs(state[0]) ** precision(0.25)
+            out[0] = math_copysign(magnitude, state[0])
+
+    @cuda.jit(device=True)
+    def operator(
+        state, parameters, drivers, base_state, t, h, a_ij, vec, out
+    ):
+        if kind == "linear":
+            out[0] = -vec[0]
+        elif kind == "root":
+            out[0] = (
+                precision(0.25)
+                * abs(state[0]) ** precision(-0.75)
+                * vec[0]
+            )
+        else:
+            out[0] = vec[0]
+
+    return {"residual": residual, "operator": operator}
+
+
+@pytest.fixture(scope="session")
+def newton_edge_solver(newton_edge_case, newton_edge_system, precision):
+    """Build the Newton solver for one edge case."""
+    case = newton_edge_case
+    linear_solver = MRLinearSolver(
+        precision=precision,
+        n=1,
+        krylov_atol=case["krylov_atol"],
+        krylov_rtol=0.0,
+        krylov_max_iters=case["krylov_max_iters"],
+    )
+    linear_solver.update(operator_apply=newton_edge_system["operator"])
+    newton = NewtonKrylov(
+        precision=precision,
+        n=1,
+        linear_solver=linear_solver,
+        newton_atol=case["newton_atol"],
+        newton_rtol=0.0,
+        newton_max_iters=case["newton_max_iters"],
+    )
+    newton.update(residual_function=newton_edge_system["residual"])
+    return newton
+
+
+@pytest.fixture(scope="session")
+def newton_edge_kernel(newton_edge_solver, precision):
+    """Compile the two-solve kernel once per parameter set."""
+    solver = newton_edge_solver.device_function
+    shared_size = max(newton_edge_solver.shared_buffer_size, 1)
+    persistent_size = max(
+        newton_edge_solver.persistent_local_buffer_size, 1
+    )
+
+    @cuda.jit
+    def kernel(states, statuses, counts):
+        parameters = cuda.local.array(1, precision)
+        drivers = cuda.local.array(1, precision)
+        base_state = cuda.local.array(1, precision)
+        counters = cuda.local.array(2, np.int32)
+        shared = cuda.shared.array(shared_size, precision)
+        persistent = cuda.local.array(persistent_size, precision)
+        for index in range(shared_size):
+            shared[index] = precision(0.0)
+        for index in range(persistent_size):
+            persistent[index] = precision(0.0)
+        parameters[0] = precision(0.0)
+        drivers[0] = precision(0.0)
+        base_state[0] = precision(0.0)
+        for solve in range(2):
+            counters[0] = np.int32(0)
+            counters[1] = np.int32(0)
+            statuses[solve] = solver(
+                states[solve : solve + 1],
+                parameters,
+                drivers,
+                precision(0.0),
+                precision(1.0),
+                precision(1.0),
+                base_state,
+                base_state,
+                shared,
+                persistent,
+                counters,
+            )
+            counts[solve] = counters[0]
+
+    return kernel
+
+
+@pytest.fixture(scope="function")
+def newton_edge_outcome(newton_edge_case, newton_edge_kernel, precision):
+    """Run two sequential solves and return finals/statuses/counts."""
+    states = cuda.to_device(
+        np.array(newton_edge_case["initials"], dtype=precision)
+    )
+    statuses = cuda.to_device(np.zeros(2, dtype=np.int32))
+    counts = cuda.to_device(np.zeros(2, dtype=np.int32))
+    newton_edge_kernel[1, 1](states, statuses, counts)
+    cuda.synchronize()
+    return (
+        states.copy_to_host(),
+        statuses.copy_to_host(),
+        counts.copy_to_host(),
+    )
 
 
 @pytest.fixture(scope="function")
@@ -309,8 +514,6 @@ def newton_solver_instance(
         newton_atol=solver_settings["newton_atol"],
         newton_rtol=solver_settings["newton_rtol"],
         newton_max_iters=solver_settings["newton_max_iters"],
-        newton_damping=solver_settings["newton_damping"],
-        newton_max_backtracks=solver_settings["newton_max_backtracks"],
     )
     solver.update(residual_function=system_setup["residual"])
     return solver
