+++ tests/test_result_codes.py
@@ -11,7 +11,6 @@ from cubie.result_codes import decode_status_codes
     "member, value",
     [
         ("SUCCESS", 0),
-        ("NEWTON_BACKTRACKING_NO_SUITABLE_STEP", 1),
         ("MAX_NEWTON_ITERATIONS_EXCEEDED", 2),
         ("MAX_LINEAR_ITERATIONS_EXCEEDED", 4),
         ("STEP_TOO_SMALL", 8),
